app.py/requirements.txt
fastapi>=0.111.0
av>=10.0.0
numpy>=1.24.0
onnxruntime-directml>=1.23.0
opencv-python-headless>=4.10.0
pandas>=2.0.0
psutil>=5.9.0
pydantic>=1.10.0,<2
pynvml>=11.4.1
pywebview>=4.4.2
PyTurboJPEG>=1.7.0
uvicorn[standard]>=0.30.0
